+++ tests/Integration/AbstractTechproductsTest.php
@@ -441,6 +441,7 @@ abstract class AbstractTechproductsTest extends TestCase
         /** @var ValueGroup $valueGroup */
         $valueGroup = $groupIterator->current();
         $this->assertSame(1, $valueGroup->getNumFound());
+        $this->assertSame(0, $valueGroup->getStart());
         $this->assertSame('A-DATA Technology Inc.', $valueGroup->getValue());
         $docIterator = $valueGroup->getIterator();
         /** @var Document $doc */
@@ -450,6 +451,7 @@ abstract class AbstractTechproductsTest extends TestCase
         $groupIterator->next();
         $valueGroup = $groupIterator->current();
         $this->assertSame(1, $valueGroup->getNumFound());
+        $this->assertSame(0, $valueGroup->getStart());
         $this->assertSame('ASUS Computer Inc.', $valueGroup->getValue());
         $docIterator = $valueGroup->getIterator();
         $doc = $docIterator->current();
@@ -458,6 +460,7 @@ abstract class AbstractTechproductsTest extends TestCase
         $groupIterator->next();
         $valueGroup = $groupIterator->current();
         $this->assertSame(1, $valueGroup->getNumFound());
+        $this->assertSame(0, $valueGroup->getStart());
         $this->assertSame('Apache Software Foundation', $valueGroup->getValue());
         $docIterator = $valueGroup->getIterator();
         $doc = $docIterator->current();
@@ -466,6 +469,7 @@ abstract class AbstractTechproductsTest extends TestCase
         $groupIterator->next();
         $valueGroup = $groupIterator->current();
         $this->assertSame(1, $valueGroup->getNumFound());
+        $this->assertSame(0, $valueGroup->getStart());
         $this->assertSame('Canon Inc.', $valueGroup->getValue());
         $docIterator = $valueGroup->getIterator();
         $doc = $docIterator->current();
@@ -474,6 +478,7 @@ abstract class AbstractTechproductsTest extends TestCase
         $groupIterator->next();
         $valueGroup = $groupIterator->current();
         $this->assertSame(2, $valueGroup->getNumFound());
+        $this->assertSame(0, $valueGroup->getStart());
         $this->assertSame('Corsair Microsystems Inc.', $valueGroup->getValue());
         $docIterator = $valueGroup->getIterator();
         $doc = $docIterator->current();
@@ -495,6 +500,8 @@ abstract class AbstractTechproductsTest extends TestCase
         /** @var QueryGroup $queryGroup */
         $queryGroup = $groupingComponentResult->getGroup('price:[0 TO 99.99]');
         $this->assertSame(5, $queryGroup->getMatches());
+        $this->assertSame(1, $queryGroup->getNumFound());
+        $this->assertSame(0, $queryGroup->getStart());
         $this->assertCount(1, $queryGroup);
         $docIterator = $queryGroup->getIterator();
         $doc = $docIterator->current();
@@ -505,6 +512,8 @@ abstract class AbstractTechproductsTest extends TestCase
 
         $queryGroup = $groupingComponentResult->getGroup('price:[100 TO *]');
         $this->assertSame(5, $queryGroup->getMatches());
+        $this->assertSame(3, $queryGroup->getNumFound());
+        $this->assertSame(0, $queryGroup->getStart());
         $this->assertCount(3, $queryGroup);
         $docIterator = $queryGroup->getIterator();
         $doc = $docIterator->current();
@@ -524,6 +533,64 @@ abstract class AbstractTechproductsTest extends TestCase
             'id' => '0579B002',
             'price' => 179.99,
         ], $doc->getFields());
+    }
+
+    /**
+     * Test fix for maxScore being returned as "NaN" when group.query doesn't match any docs.
+     *
+     * Skipped for SolrCloud because maxScore is included in distributed search results even if score is not requested (SOLR-6612).
+     * This makes the test fail on SolrCloud for queries that don't fetch a score and thus aren't affected by SOLR-13839.
+     *
+     * @group skip_for_solr_cloud
+     *
+     * @see https://issues.apache.org/jira/browse/SOLR-13839
+     * @see https://issues.apache.org/jira/browse/SOLR-6612
+     */
+    public function testGroupingComponentFixForSolr13839()
+    {
+        self::$client->registerQueryType('grouping', '\Solarium\Tests\Integration\GroupingTestQuery');
+        /** @var GroupingTestQuery $select */
+        $select = self::$client->createQuery('grouping');
+        // without score in the fl parameter, result groups don't have a maxScore
+        $select->setFields('id');
+        $grouping = $select->getGrouping();
+        $grouping->addQueries([
+            'cat:memory',
+            'cat:no-such-cat',
+        ]);
+        $result = self::$client->select($select);
+        $groupingComponentResult = $result->getComponent(ComponentAwareQueryInterface::COMPONENT_GROUPING);
+
+        /** @var QueryGroup $queryGroup */
+        $queryGroup = $groupingComponentResult->getGroup('cat:memory');
+        $this->assertSame(32, $queryGroup->getMatches());
+        $this->assertSame(3, $queryGroup->getNumFound());
+        $this->assertNull($queryGroup->getMaximumScore());
+
+        $queryGroup = $groupingComponentResult->getGroup('cat:no-such-cat');
+        $this->assertSame(32, $queryGroup->getMatches());
+        $this->assertSame(0, $queryGroup->getNumFound());
+        $this->assertNull($queryGroup->getMaximumScore());
+
+        // with score in the fl parameter, result groups have a maxScore
+        $select->setFields('id,score');
+        $grouping = $select->getGrouping();
+        $grouping->addQueries([
+            'cat:memory',
+            'cat:no-such-cat',
+        ]);
+        $result = self::$client->select($select);
+        $groupingComponentResult = $result->getComponent(ComponentAwareQueryInterface::COMPONENT_GROUPING);
+
+        $queryGroup = $groupingComponentResult->getGroup('cat:memory');
+        $this->assertSame(32, $queryGroup->getMatches());
+        $this->assertSame(3, $queryGroup->getNumFound());
+        $this->assertNotNull($queryGroup->getMaximumScore());
+
+        $queryGroup = $groupingComponentResult->getGroup('cat:no-such-cat');
+        $this->assertSame(32, $queryGroup->getMatches());
+        $this->assertSame(0, $queryGroup->getNumFound());
+        $this->assertNull($queryGroup->getMaximumScore());
     }
 
     public function testQueryElevation()